+++ test/vector_prop_test.exs
@@ -187,8 +187,8 @@ defmodule A.Vector.PropTest do
       assert slice_2 === Enum.slice(vector, i1..i2)
       assert A.Vector.new(slice_2) === A.Vector.slice(vector, i1..i2)
 
-      assert Enum.take(list, i1) |> A.Vector.new() == A.Vector.take(vector, i1)
-      assert Enum.drop(list, i1) |> A.Vector.new() == A.Vector.drop(vector, i1)
+      assert Enum.take(list, i1) |> A.Vector.new() === A.Vector.take(vector, i1)
+      assert Enum.drop(list, i1) |> A.Vector.new() === A.Vector.drop(vector, i1)
 
       deleted_list = List.delete_at(list, i1)
       assert A.Vector.new(deleted_list) == A.Vector.delete_at(vector, i1)
@@ -207,6 +207,9 @@ defmodule A.Vector.PropTest do
       filtered_list = Enum.filter(list, &hash_multiple_of_2/1)
       filtered_vector = A.Vector.filter(vector, &hash_multiple_of_2/1)
       assert A.Vector.new(filtered_list) === filtered_vector
+
+      index_list = Enum.with_index(list, i1)
+      assert A.Vector.new(index_list) === A.Vector.with_index(vector, i1)
 
       assert Enum.any?(list) === A.Vector.any?(vector)
       assert Enum.all?(list) === A.Vector.all?(vector)