+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Dev
 
+### Enhancements
+
+* Add `A.Vector.with_index/2`
+
 ## v0.4.3 (2021-01-12)
 
 This release is mostly focused on vector slicing and performance ⚡️